requirements.txt
mnemonic==0.21
pycryptodome==3.23.0
